requirements.txt
Django==2.2
django-model-utils==4.0.0
djangorestframework==3.12.1
flake8==3.8.4
flake8-django==1.1.1
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.4.1
